renamed tp "propvision"
